pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aumai-ragoptimizer"
version = "0.1.0"
description = "Benchmark and optimize RAG pipeline configurations"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "AumAI Contributors", email = "dev@aumai.dev" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "hypothesis>=6.0",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.urls]
Homepage = "https://github.com/aumai/aumai-ragoptimizer"
Documentation = "https://docs.aumai.dev/aumai-ragoptimizer"
Repository = "https://github.com/aumai/aumai-ragoptimizer"
Issues = "https://github.com/aumai/aumai-ragoptimizer/issues"

[project.scripts]
aumai-ragoptimizer = "aumai_ragoptimizer.cli:main"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "ANN", "S", "B"]

[tool.mypy]
strict = true
python_version = "3.11"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
